Moved some of the sound functions into the .cpp file for better organization between header and source file
